% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_composantes_valorisees_format.R
\name{ana_composantes_valorisees_format}
\alias{ana_composantes_valorisees_format}
\title{Transformation des données sur les composantes valorisées}
\usage{
ana_composantes_valorisees_format()
}
\description{
Transformation des données sur les composantes valorisées
}
\details{
Cette fonction effectue transforme les données sur les composantes valorisées pour les préparer pour l'évaluation des effets cumualatifs
}
\keyword{composantes}
\keyword{transformation}
\keyword{valorisées}
