% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_composantes_valorisees_raw.R
\name{ana_composantes_valorisees_raw}
\alias{ana_composantes_valorisees_raw}
\title{Composantes valorisées}
\usage{
ana_composantes_valorisees_raw()
}
\description{
Base de données avec l'ensemble des composantes valorisées brutes
}
\details{
Cette fonction effectue une partie des analyses du projet d'évaluation des effets cumulatifs
}
\keyword{composantes}
\keyword{valorisées}
