% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_cumulative_composantes_valorisees.R
\name{ana_cumulative_composantes_valorisees}
\alias{ana_cumulative_composantes_valorisees}
\title{Empreinte cumulée des composantes valorisées}
\usage{
ana_cumulative_composantes_valorisees()
}
\description{
Évaluation de l'empreinte cumulée des composantes valorisées considérés
}
\details{
Cette fonction effectue une partie des analyses du projet d'évaluation des effets cumulatifs
}
\keyword{composantes}
\keyword{cumulée}
\keyword{empreinte}
\keyword{valorisées}
