% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_cumulative_effects.R
\name{ana_cumulative_effects}
\alias{ana_cumulative_effects}
\title{Effets cumulatifs}
\usage{
ana_cumulative_effects()
}
\description{
Évaluation des effets cumulatives des stresseurs environnementaux sur les  composantes valorisées
}
\details{
Cette fonction effectue une partie des analyses du projet d'évaluation des effets cumulatifs
}
\keyword{composantes}
\keyword{cumulatifs}
\keyword{effets}
\keyword{stresseurs}
\keyword{valorisées}
