% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_cumulative_effects_region_km2.R
\name{ana_cumulative_effects_region_km2}
\alias{ana_cumulative_effects_region_km2}
\title{Effets cumulatifs par km2 sur les régions administratives du Saint-Laurent}
\usage{
ana_cumulative_effects_region_km2()
}
\description{
Évaluation des effets cumulatifs par km^2 des stresseurs environnementaux sur les composantes valorisées dans les régions administratives du Saint-Laurent
}
\details{
Cette fonction effectue une partie des analyses du projet d'évaluation des effets cumulatifs
}
\keyword{composantes}
\keyword{cumulatifs}
\keyword{effets}
\keyword{stresseurs}
\keyword{valorisées}
