% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_cumulative_exposure.R
\name{ana_cumulative_exposure}
\alias{ana_cumulative_exposure}
\title{Exposition cumulée}
\usage{
ana_cumulative_exposure()
}
\description{
Évaluation de l'exposition cumulée des composantes valorisées aux stresseurs environnementaux
}
\details{
Cette fonction effectue une partie des analyses du projet d'évaluation des effets cumulatifs
}
\keyword{composantes}
\keyword{cumulée}
\keyword{exposition}
\keyword{stresseurs}
\keyword{valorisées}
