% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_stresseurs_format.R
\name{ana_stresseurs_format}
\alias{ana_stresseurs_format}
\title{Transformation des données sur les stresseurs environnementaux}
\usage{
ana_stresseurs_format()
}
\description{
Transformation des données sur les stresseurs environnementaux
}
\details{
Cette fonction effectue transforme les données sur les stresseurs environnementaux pour les préparer pour l'évaluation des effets cumualatifs
}
\keyword{stresseurs}
\keyword{transformation}
