% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_vulnerability.R
\name{ana_vulnerability}
\alias{ana_vulnerability}
\title{Identification de la vulnérabilité des composantes valorisées aux stresseurs environnementaux}
\usage{
ana_vulnerability()
}
\description{
Identification de la vulnérabilité des composantes valorisées aux stresseurs environnementaux
}
\details{
Cette fonction effectue formate la base de donnée de vulnérabilité des composantes valorisées aux stresseurs environnementaux considérés
}
\keyword{vulnérabilité}
