% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_basemap.R
\name{basemap}
\alias{basemap}
\title{Load basemap data}
\usage{
basemap(data_name)
}
\arguments{
\item{data_name}{\code{character} name of basemap data to load in R session}
}
\description{
Load basemap data
}
