% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_load.R
\name{ceanav_load}
\alias{ceanav_load}
\title{Load formatted data available in \verb{./data/data-format/}}
\usage{
ceanav_load(data_id)
}
\arguments{
\item{data_id}{\code{character} id of data to import in R session with format \code{dataXXXX}}
}
\description{
The function imports data available in \verb{./data/data-format/} regardless of its format.
}
\details{
This function is used to avoid putting all formatted dataset in lazy load, yet still make it easy for them to be loaded. The user simply provides the id of the dataset and the function imports it in the R session
}
\examples{
ceanav_load('data0001')
}
\keyword{contact}
\keyword{metadata}
