% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_plot.R
\name{ceanav_plot}
\alias{ceanav_plot}
\alias{ceanav_plot.sf}
\title{plot ceanav data}
\usage{
ceanav_plot(dat, ...)

\method{ceanav_plot}{sf}(dat, ...)
}
\arguments{
\item{dat}{object of class sf, stars}

\item{...}{further specifications, see \link{plot} and details.}
}
\description{
base plot functions for ceanav project
}
