% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_metadata.R
\name{ceanav_update_metadata}
\alias{ceanav_update_metadata}
\title{Update metadata and contacts}
\usage{
ceanav_update_metadata(type)
}
\arguments{
\item{type}{type of metadate, one of \code{metadata} or \code{contact}}
}
\description{
The function updates the existing metadata file based on individual metadata files in \code{data-metadata}
}
\details{
These functions take '.yml' files in specified folder to generate an individual table containing information from all \code{.yml} files
}
\examples{
# Metadata
ceanav_metadata('metadata')

# Contacts
ceanav_metadata('contact')
}
\keyword{contact}
\keyword{metadata}
