% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_cumulative_effect.R
\name{cumulativeEffects}
\alias{cumulativeEffects}
\title{Cumulative effects}
\usage{
cumulativeEffects(
  stress,
  valued,
  vulnerability,
  output_name = "cumulative_effects",
  individual_cea = TRUE
)
}
\arguments{
\item{stress}{matrix with stressors intensity}

\item{valued}{matrix with presence or amount of valued components}

\item{vulnerability}{matrix with vulnerability of valued components to stressors. Names used in stress and valued matrices must be the same as that found in vulnerability matrix}

\item{output_name}{character, name of the output file}

\item{individual_cea}{logical, if TRUE function exports all individual effects}
}
\description{
Function to evaluate cumulative effects of stressors on valued components
}
\details{

}
\keyword{cumulative}
\keyword{effects}
