% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_cumulative_footprint.R
\name{cumulativeFootprint}
\alias{cumulativeFootprint}
\title{Cumulative footprint}
\usage{
cumulativeFootprint(dat, normaliser = FALSE)
}
\arguments{
\item{dat}{data.frame with stressors intensity or valued components}

\item{normaliser}{logical, whether to normalize footprint}
}
\description{
Function to evaluate cumulative footprint of stressors and valued components
}
\details{

}
\keyword{cumulative}
\keyword{footprint}
