% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_berge.R
\name{cv_berge}
\alias{cv_berge}
\title{Berges: Données sur l'intégrité des berges}
\usage{
cv_berge()
}
\description{
Intégration des données utilisées pour caractériser la composante valorisée d'intégrité des berges
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{berge}
\keyword{composante}
\keyword{valorisée}
