% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_habitat.R
\name{cv_habitats}
\alias{cv_habitats}
\title{Habitats: Integrated habitats dataset}
\usage{
cv_habitats()
}
\description{
Intégration des données utilisées pour caractériser la composante valorisée d'habitats
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{humide}
\keyword{milieu}
\keyword{valorisée}
\keyword{zostère}
