% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_cv_mammiferes_marins.R
\name{cv_mammiferes_marins}
\alias{cv_mammiferes_marins}
\title{Mammifères marins: Données sur la présence de mammifères marins}
\usage{
cv_mammiferes_marins()
}
\description{
Intégration des données utilisées pour caractériser la composante valorisée sur les mammifères marins
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{mammifères}
\keyword{marins}
\keyword{valorisée}
