% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_cv_site.R
\name{cv_site}
\alias{cv_site}
\title{Sites d’intérêt culturels, patrimoniaux et archéologiques}
\usage{
cv_site()
}
\description{
Intégration des données utilisées pour caractériser la composante valorisée des sites d’intérêt culturels, patrimoniaux et archéologiques
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{archéologiques}
\keyword{composante}
\keyword{culturels,}
\keyword{d’intérêt}
\keyword{et}
\keyword{patrimoniaux}
\keyword{sites}
\keyword{valorisée}
