% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0001-zostere.R
\name{data0001_zostere}
\alias{data0001_zostere}
\title{Data 0001 : Inventaire zostère}
\source{
https://ouvert.canada.ca/data/fr/dataset/c9ab948f-5009-4dbc-9129-2f6e373f17f6
}
\usage{
data0001_zostere()
}
\description{
Inventaire de la zostère marine dans la Baie James, la Baie des Chaleurs, l'estuaire et le golfe du Saint-Laurent
}
\details{
Cette fonction télécharge les données et les préparent pour les analyses
}
\keyword{composante}
\keyword{habitat}
\keyword{valorisée}
\keyword{zostère}
