% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Tableau récapitulatif des données}
\usage{
data_summary()
}
\description{
Fonction utilisée pour générer un tableau récapitulatif des données présentées dans le rapport.
}
\details{
Cette fonction génère un tableau récapitulatif des données considérées pour l'évaluation
}
\examples{
fnc_data_summary()
}
\keyword{récapitulatif}
\keyword{tableau}
