% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_diffusive_model.R
\name{diffusive}
\alias{diffusive}
\title{Diffusive model}
\usage{
diffusive(dat, field, threshold, globalmaximum, decay, distance, increment)
}
\arguments{
\item{dat}{sf object with points characterizing stressor intensity}

\item{field}{character, name of field containing stresseor intensity}

\item{threshold}{numeric, minimum threshold in percent of the global maximum at which to stop the diffusive model}

\item{globalmaximum}{global maximum value of the stressor in the study area}

\item{decay}{percent by which the value of the stressor is reduced when diffusing to the adjacent cells}

\item{distance}{numeric, distance beyond which the stressor is considered to have no influence. Units are in the units of the spatial object provided.}

\item{increment}{numeric, distance between successive steps in the model. Units are in the units of the spatial object provided.}
}
\description{
Function to evaluate the influence area of a stressor using a passive diffusive model
}
\details{

}
\keyword{cumulative}
\keyword{footprint}
