% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_cumulative_composantes_valorisees.R
\name{fig_cumulative_composantes_valorisees}
\alias{fig_cumulative_composantes_valorisees}
\title{Figures composantes valorisées cumulés}
\usage{
fig_cumulative_composantes_valorisees(lang = "fr")
}
\description{
Fonctions pour générer des figures pour les composantes_valorisees cumulés
}
\examples{
fig_cumulative_composantes_valorisees()
}
\keyword{figure}
