% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_cumulative_effects.R
\name{fig_cumulative_effects}
\alias{fig_cumulative_effects}
\title{Figures effets cumulatifs0}
\usage{
fig_cumulative_effects()
}
\description{
Fonctions pour générer des figures pour les effets cumulatifs
}
\examples{
fig_cumulative_effects()
}
\keyword{figure}
