% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_cumulative_exposure.R
\name{fig_cumulative_exposure}
\alias{fig_cumulative_exposure}
\title{Figures exposition cumul/e}
\usage{
fig_cumulative_exposure()
}
\description{
Fonctions pour générer des figures pour l'exposition cumulée
}
\examples{
fig_cumulative_exposure()
}
\keyword{figure}
