% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_cumulative_hotspots.R
\name{fig_cumulative_hotspots}
\alias{fig_cumulative_hotspots}
\title{Figures hotspots cumulés}
\usage{
fig_cumulative_hotspots(lang = "fr")
}
\description{
Fonctions pour générer des figures pour les hotspots cumulés
}
\examples{
fig_cumulative_hotspots()
}
\keyword{figure}
