% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_metanetwork.R
\name{fig_metanetwork}
\alias{fig_metanetwork}
\title{Figure metanetwork}
\usage{
fig_metanetwork()
}
\description{
Figure metanetwork
}
\examples{
# Figure for specific dataset
fig_metanetwork()
}
\keyword{figure}
