% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_metanetwork_en.R
\name{fig_metanetwork_en}
\alias{fig_metanetwork_en}
\title{Figure metanetwork}
\usage{
fig_metanetwork_en()
}
\description{
Figure metanetwork
}
\examples{
# Figure for specific dataset
fig_metanetwork()
}
\keyword{figure}
