% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_navigation_exploration.R
\name{fig_navigation_exploration}
\alias{fig_navigation_exploration}
\title{Figures données de navigation pour exploration}
\usage{
fig_navigation_exploration()
}
\description{
Fonctions pour générer des figures pour les données de navigation AIS
}
\keyword{figure}
