% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_region_cea_km2.R
\name{fig_region_cea_km2}
\alias{fig_region_cea_km2}
\title{Figure effects cumulatifs par région administrative par km2}
\usage{
fig_region_cea_km2(lang = "fr")
}
\description{
Figure effects cumulatifs par région administrative par km2
}
\examples{
# Figure for specific dataset
fig_region_cea_km2()
}
\keyword{figure}
