% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_regional_contribution.R
\name{fig_regional_contribution}
\alias{fig_regional_contribution}
\title{Figure contribution régionale par km2}
\usage{
fig_regional_contribution()
}
\description{
Contribution régionale des effets cumulatifs par km2 pour chaque stresseur sur chaque composante valorisée
}
\examples{
# Figure for specific dataset
fig_regional_contribution()
}
\keyword{figure}
