% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_regional_contribution_en.R
\name{fig_regional_contribution_en}
\alias{fig_regional_contribution_en}
\title{Figure contribution régionale par km2 (english version)}
\usage{
fig_regional_contribution_en()
}
\description{
Contribution régionale des effets cumulatifs par km2 pour chaque stresseur sur chaque composante valorisée
}
\examples{
# Figure for specific dataset
fig_regional_contribution()
}
\keyword{figure}
