% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_heatmap.R
\name{fnc_heatmap}
\alias{fnc_heatmap}
\title{Function to generate a heatmap with given data, and rows and columns names}
\usage{
fnc_heatmap(dat, colums, rows, output)
}
\description{
Function to generate a heatmap with given data, and rows and columns names
}
\keyword{figure}
