% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0000_aoi.R
\name{get_aoi}
\alias{get_aoi}
\title{Data 0000 : Zone d'intérêt et grilles d'étude}
\usage{
get_aoi()
}
\description{
Délimitation de la zone d'intérêt et grilles utilisées pour l'étude
}
\details{
Cette fonction télécharge les données
}
\keyword{d'intérêt}
\keyword{d'étude}
\keyword{grille}
\keyword{zone}
