% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0000_basemap.R
\name{get_basemap}
\alias{get_basemap}
\title{Download and format basemap data used for figures}
\usage{
get_basemap()
}
\description{
Accesses and formats spatial data that I use frequently for static mapping
}
