% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0001-zostere.R
\name{get_data0001}
\alias{get_data0001}
\title{Data 0001 : Inventaire zostère}
\source{
https://ouvert.canada.ca/data/fr/dataset/c9ab948f-5009-4dbc-9129-2f6e373f17f6
}
\usage{
get_data0001()
}
\description{
Inventaire de la zostère marine dans la Baie James, la Baie des Chaleurs, l'estuaire et le golfe du Saint-Laurent provenant d'une revue de littérature de documents entre 1987 et 2009
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{valorisée}
\keyword{zostère}
