% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0002-zostere.R
\name{get_data0002}
\alias{get_data0002}
\title{Data 0002 : Zostère Pointe au père}
\source{
https://catalogue.ogsl.ca/fr/dataset/3d0c057b-ab39-4a7b-8809-4367d5028c11
}
\usage{
get_data0002()
}
\description{
Données d'un projet de restauration de l'herbier de zostère marine dans l'anse de Point-au-Père
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{valorisée}
\keyword{zostère}
