% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0004-milieu_humide.R
\name{get_data0004}
\alias{get_data0004}
\title{Data 0004 : Milieux humides Lac St-Pierre}
\source{
https://ouvert.canada.ca/data/fr/dataset/9c52df44-7a34-4a73-a92a-5e3b20de6c73
}
\usage{
get_data0004()
}
\description{
Données de terrain pour la cartographie des milieux humides du lac Saint-Pierre (fleuve Saint-Laurent), 2012
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{humide}
\keyword{lac}
\keyword{milieu}
\keyword{st-pierre}
\keyword{valorisée}
