% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0006-milieu_humide.R
\name{get_data0006}
\alias{get_data0006}
\title{Data 0006 : Milieux humides Saint-Laurent}
\source{
https://catalogue.ogsl.ca/fr/dataset/0a232214-05cc-438a-b914-6a8b53ac184e
}
\usage{
get_data0006()
}
\description{
Atlas des milieux côtiers d’intérêt pour la conservation dans l’estuaire et le golfe du Saint-Laurent
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{humide}
\keyword{milieu}
\keyword{valorisée}
