% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0008-milieu_sableux.R
\name{get_data0008}
\alias{get_data0008}
\title{Data 0008 : Milieu sableux Saint-Laurent}
\source{
https://catalogue.ogsl.ca/fr/dataset/0a232214-05cc-438a-b914-6a8b53ac184e
}
\usage{
get_data0008()
}
\description{
Atlas des milieux côtiers d’intérêt pour la conservation dans l’estuaire et le golfe du Saint-Laurent
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{milieu}
\keyword{sableux}
\keyword{valorisée}
