% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0010-frayere.R
\name{get_data0010}
\alias{get_data0010}
\title{Data 0010 : Frayères}
\source{
Marc Mingelbier, Ministère des Forêts, de la Faune et des Parcs
}
\usage{
get_data0010()
}
\description{
Identification des frayères dans le Saint-Laurent fluvial
}
\details{
Cette fonction formatte les données
}
\keyword{composante}
\keyword{frayère}
\keyword{habitat}
\keyword{valorisée}
