% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0011-espece_statut.R
\name{get_data0011}
\alias{get_data0011}
\title{Data 0011 : Espèces à statut}
\source{
Marc Mingelbier, Ministère des Forêts, de la Faune et des Parcs
}
\usage{
get_data0011()
}
\description{
Zones d'importance pour les espèces à statut dans le Saint-Laurent fluvial
}
\details{
Cette fonction formatte les données
}
\keyword{composante}
\keyword{espèce}
\keyword{habitat}
\keyword{statut}
\keyword{valorisée}
\keyword{à}
