% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0012-milieu_cotier.R
\name{get_data0012}
\alias{get_data0012}
\title{Data 0012 : Classification rivages côtiers}
\source{
https://ouvert.canada.ca/data/fr/dataset/ba580518-59e8-4d1c-b3ef-41d2658e6965
}
\usage{
get_data0012()
}
\description{
Classification des Rivages du Québec - Fleuve Saint-Laurent
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{côtier}
\keyword{habitat}
\keyword{milieu}
\keyword{rivage}
\keyword{valorisée}
