% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0013-zone_inondable.R
\name{get_data0013}
\alias{get_data0013}
\title{Data 0013 : Zones inondables MRC}
\source{
https://www.donneesquebec.ca/recherche/fr/dataset/0d9de0d6-9873-4a8c-adc7-0e94d51b3fa0
}
\usage{
get_data0013()
}
\description{
Grille de présence de zone inondable identifiée par les MRC
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{inondable}
\keyword{valorisée}
\keyword{zone}
