% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0014-zone_inondable.R
\name{get_data0014}
\alias{get_data0014}
\title{Data 0014 : Zones inondables BDZI}
\source{
https://www.donneesquebec.ca/recherche/fr/dataset/3ac8ddff-fe0a-4a7a-8393-d5938e8f35e5
}
\usage{
get_data0014()
}
\description{
Base de données des zones à risque d'inondation (BDZI)
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{inondable}
\keyword{valorisée}
\keyword{zone}
