% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0015-ancrage.R
\name{get_data0015}
\alias{get_data0015}
\title{Data 0015 : Ancrages}
\source{
INNAV
}
\usage{
get_data0015()
}
\description{
Distribution des sites d'ancrages de SIGTM-INNAV dans l'Est du Canada
}
\details{
Cette fonction formatte les données
}
\keyword{ancrage}
\keyword{stresseurs}
