% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0016-deversement.R
\name{get_data0016}
\alias{get_data0016}
\title{Data 0016 : Déversements accidentels}
\source{
Garde côtière canadienne, Direction des Interventions, Intervention environnementale, région du Centre. Base de données SGI-WEB -2007-2020.
}
\usage{
get_data0016()
}
\description{
Distribution des sites de déversements accidentels répertoriés dans le Saint-Laurent
}
\details{
Cette fonction formatte les données
}
\keyword{accidentel}
\keyword{déversement}
\keyword{stresseurs}
