% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0017-berge.R
\name{get_data0017}
\alias{get_data0017}
\title{Data 0017 : Caractérisation des berges}
\source{
https://www.donneesquebec.ca/recherche/dataset/caracterisation-des-berges-et-analyse-de-l-evolution-des-facteurs-hydro-climatiques
}
\usage{
get_data0017()
}
\description{
Caractérisation des berges de la partie fluviale du Saint-Laurent et analyse de l'évolution des facteurs hydro-climatiques influençant les aléas d'érosion et d'inondation
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{berge}
\keyword{composante}
\keyword{habitat}
\keyword{valorisée}
