% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0018-dragage.R
\name{get_data0018}
\alias{get_data0018}
\title{Data 0018 : Chenal de navigation et dragage}
\source{
Garde côtière canadienne
}
\usage{
get_data0018()
}
\description{
Chenal de navigation entretenu par des activités de dragage au sein du Saint-Laurent et localisation des sites de dépôts utilisés
}
\details{
Cette fonction formatte les données
}
\keyword{chenal}
\keyword{de}
\keyword{dragage}
\keyword{navigation}
\keyword{stresseurs}
