% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0020-navigation.R
\name{get_data0020}
\alias{get_data0020}
\title{Data 0020 : Index des navires données AIS (id: 0021)}
\source{
Transports Canada
}
\usage{
get_data0020()
}
\description{
Base de données contenant les informations des navires pour lesquels des données AIS satellitaire (id: 0021) sont disponibles
}
\details{
Cette fonction formatte les données
}
\keyword{navigation}
\keyword{stresseurs}
