% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0021-navigation.R
\name{get_data0021}
\alias{get_data0021}
\title{Data 0021 : Données AIS satellite de navigation}
\source{
Transports Canada, Maerospace Corporation
}
\usage{
get_data0021()
}
\description{
Satellite navigation data from automatic identification system (AIS) in the St. Lawrence estuary from 2017 to 2019.
}
\details{
Cette fonction formatte les données
}
\keyword{navigation}
\keyword{stresseurs}
