% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0022-peche_aghamm.R
\name{get_data0022}
\alias{get_data0022}
\title{Data 0022 : Pêches commerciales AGHAMM}
\source{
https://catalogue.ogsl.ca/en/dataset/bbe35312-987a-42d7-ae2f-41c1a0715dc3
}
\usage{
get_data0022()
}
\description{
Mi'gmaq and Maliseet commercial fisheries of the communities of Gesgapegiag, Gespeg and Viger
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{commerciale}
\keyword{composante}
\keyword{d'importance}
\keyword{pêche}
\keyword{sites}
\keyword{valorisée}
