% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0023-saumon_aghamm.R
\name{get_data0023}
\alias{get_data0023}
\title{Data 0023 : Saumon Atlantique AGHAMM}
\source{
Atlas AGHAMM
}
\usage{
get_data0023()
}
\description{
Important sites for Atlantic salmon for the communities of Gesgapegiag, Gespeg and Viger
}
\details{
Cette fonction formatte les données
}
\keyword{Atlantique}
\keyword{composante}
\keyword{d'importance}
\keyword{saumon}
\keyword{sites}
\keyword{valorisée}
