% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0024-essipit.R
\name{get_data0024}
\alias{get_data0024}
\title{Data 0024 : Sites d'importance communauté Essipit}
\source{
Bouchard D., Tremblay P. (communication personnelle)  Collection de sites d'intérêt sociaux, culturels, archéologiques et de pratiques Innu-Aitun issue de l'intégration d'ouvrages divers. Secteur Territoire et Consultation, Conseil de Bande Essipit.
}
\usage{
get_data0024()
}
\description{
Collection de sites d'intérêt sociaux, culturels, archéologiques et de pratiques Innu-Aitun issue de l'intégration d'ouvrages divers
}
\details{
Cette fonction formatte les données
}
\keyword{Essipit}
\keyword{composante}
\keyword{d'importance}
\keyword{sites}
\keyword{valorisée}
