% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0025-wolastoqiyik_wahsipekuk.R
\name{get_data0025}
\alias{get_data0025}
\title{Data 0025 : Communauté Wolastoqiyik Wahsipekuk pêche homard}
\source{
Première Nation Wolastoqiyik Wahsipekuk (2021) Localisation des sites d'importance pour la pêche aux homards dans l'estuaire du Saint-Laurent pour l'année 2019.
}
\usage{
get_data0025()
}
\description{
Localisation des sites d'importance pour la pêche aux homards dans l'estuaire du Saint-Laurent pour l'année 2019
}
\details{
Cette fonction formatte les données
}
\keyword{Américain}
\keyword{Wahsipekuk}
\keyword{Wolastoqiyik}
\keyword{composante}
\keyword{d'importance}
\keyword{homard}
\keyword{sites}
\keyword{valorisée}
