% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0026-wolastoqiyik_wahsipekuk.R
\name{get_data0026}
\alias{get_data0026}
\title{Data 0026 : Communauté Wolastoqiyik Wahsipekuk pêche oursin vert}
\source{
Première Nation Wolastoqiyik Wahsipekuk (2021) Localisation des sites d'importance pour la pêche à l'oursin vert dans l'estuaire du Saint-Laurent
}
\usage{
get_data0026()
}
\description{
Localisation des sites d'importance pour la pêche à l'oursin vert dans l'estuaire du Saint-Laurent.
}
\details{
Cette fonction formatte les données
}
\keyword{Wahsipekuk}
\keyword{Wolastoqiyik}
\keyword{composante}
\keyword{d'importance}
\keyword{oursin}
\keyword{sites}
\keyword{valorisée}
\keyword{vert}
