% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0027-mammiferes_marins.R
\name{get_data0027}
\alias{get_data0027}
\title{Data 0027 :}
\source{
Le WWF-Canada et le Réseau d’observation de mammifères marins. Données cartographiques : Navires et baleines de l’Atlantique Nord-Ouest : Guide à l’intention de l’industrie maritime, 2e édition. 2021. Données disponibles sur le site https://www.navigationbaleines.ca/fr/accueil/
}
\usage{
get_data0027()
}
\description{
Data 0027 :
}
\details{
Cette fonction formatte les données
}
\keyword{composante}
\keyword{mammifères}
\keyword{marins}
\keyword{valorisée}
