% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0028-navigation_observation.R
\name{get_data0028}
\alias{get_data0028}
\title{Data 0028 : Observation en mer de mammifères marins}
\source{
Turgeon, S., 2019. Portrait de la navigation dans le parc marin du Saguenay–Saint-Laurent - 2017. Parcs Canada, 59 pages + annexes
}
\usage{
get_data0028()
}
\description{
Activité d'observation en mer de mammifères marins sur des embarcations avec permis de classe 1 en 2017
}
\details{
Cette fonction formatte les données
}
\keyword{en}
\keyword{mer}
\keyword{navigation}
\keyword{observation}
\keyword{stresseurs}
