% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0030-mesures_conservation.R
\name{get_data0030}
\alias{get_data0030}
\title{Data 0030 : Mesures spatiales de gestion parc marin}
\source{
Agence Parcs Canada (2021) Limites des différentes mesures spatiales de gestion en place au parc marin
}
\usage{
get_data0030()
}
\description{
Limites des différentes mesures spatiales de gestion en place au parc marin
}
\details{
Cette fonction formatte les données
}
\keyword{conservation}
\keyword{de}
\keyword{mesure}
