% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0032-kahnawake.R
\name{get_data0032}
\alias{get_data0032}
\title{Data 0032 : Kahnawà:ke - Local input on valued components}
\source{
Mohawk Council of Kahnawà:ke (2021) Kahnawà:ke – Local input on valued components identified through the Cumulative Effects of Marine Shipping initiative for the St. Lawrence River
}
\usage{
get_data0032()
}
\description{
Kahnawà:ke – Local input on valued components identified through the Cumulative Effects of Marine Shipping initiative for the St. Lawrence River
}
\details{
Cette fonction formatte les données
}
\keyword{archéologique}
\keyword{chasse}
\keyword{composante}
\keyword{culturel}
\keyword{d'importance}
\keyword{kahnawà:ke}
\keyword{pêche}
\keyword{site}
\keyword{valorisée}
\keyword{végétation}
