% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0033-peche_commerciale.R,
%   R/dataXXXX-conservation.R
\name{get_data0033}
\alias{get_data0033}
\title{Data 0033 : Pêche commerciale}
\source{
Fisheries and Oceans Canada (2021). Departement of Fisheries and Oceans Canada’s Fisheries and Oceans Canada Zonal Interchange File Format (ZIFF) data. A compilation of landing data from logbook data between 2000 and 2020. \link{Data accessed: 2021-06-11}".

https://www.donneesquebec.ca/recherche/dataset/atlas-des-territoires-interet-conservation-btsl#
}
\usage{
get_data0033()

get_data0033()
}
\description{
A compilation of landing data from Zonal Interchange File Format (ZIFF) data between 2000 and 2020

Atlas des territoires d’intérêt pour la conservation dans les Basses-terres du Saint-Laurent
}
\details{
Cette fonction formatte les données

Cette fonction télécharge et formatte les données
}
\keyword{commerciale}
\keyword{composante}
\keyword{conservation}
\keyword{d'impoetance}
\keyword{d'intérêt}
\keyword{la}
\keyword{pour}
\keyword{pêche}
\keyword{site}
\keyword{stresseurs}
\keyword{territoire}
\keyword{valorisée}
