% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0034-peche_commerciale.R
\name{get_data0034}
\alias{get_data0034}
\title{Data 0034 : Index des engins de pêche pour la pêche commerciale (id: 0033)}
\source{
Fisheries and Oceans Canada (2021). Index of fishing gears in departement of Fisheries and Oceans Canada’s Fisheries and Oceans Canada Zonal Interchange File Format (ZIFF) data. A compilation of landing data from logbook data between 2000 and 2020. \link{Data accessed 2021-07-15}
}
\usage{
get_data0034()
}
\description{
Index des engin pour les données de from Zonal Interchange File Format (ZIFF) entre 2000 et 2020
}
\details{
Cette fonction formatte les données
}
\keyword{commerciale}
\keyword{pêche}
\keyword{stresseurs}
