% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0035-peche_commerciale.R
\name{get_data0035}
\alias{get_data0035}
\title{Data 0035 : Index des espèces pour la pêche commerciale (id: 0033)}
\source{
Fisheries and Oceans Canada (2021). Index of species in departement of Fisheries and Oceans Canada’s Fisheries and Oceans Canada Zonal Interchange File Format (ZIFF) data. A compilation of landing data from logbook data between 2000 and 2020. \link{Data accessed 2021-07-15}
}
\usage{
get_data0035()
}
\description{
Index des espèces pour les données de from Zonal Interchange File Format (ZIFF) entre 2000 et 2020
}
\details{
Cette fonction formatte les données
}
\keyword{commerciale}
\keyword{pêche}
\keyword{stresseurs}
