% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0036-habitat_faunique.R
\name{get_data0036}
\alias{get_data0036}
\title{Data 0036 : Habitats fauniques}
\source{
https://www.donneesquebec.ca/recherche/dataset/habitats-fauniques
}
\usage{
get_data0036()
}
\description{
La géodatabase (GDB) des habitats fauniques (HAFA) contient les données des 11 habitats fauniques légaux situés sur les terres du domaine de l’État qui respectent la définition du Règlement sur les habitats fauniques (RHF). On trouve aussi les HAFA situés sur terres mixtes et privées à titre informationnel.
}
\details{
This function loads and formats the data
}
\keyword{composantes}
\keyword{fauniques}
\keyword{habitats}
\keyword{valorisées}
