% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0037-habitat_floristique.R
\name{get_data0037}
\alias{get_data0037}
\title{Data 0037 : Habitats floristiques}
\source{
https://www.donneesquebec.ca/recherche/dataset/habitats-floristiques
}
\usage{
get_data0037()
}
\description{
Un habitat floristique est une aire protégée qui abrite au moins une espèce floristique désignée menacée ou vulnérable et qui est identifié à l’article 7 du Règlement sur les espèces floristiques menacées ou vulnérables et leurs habitats.
}
\details{
This function loads and formats the data
}
\keyword{}
