% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0038-aires_protogees.R
\name{get_data0038}
\alias{get_data0038}
\title{Data 0040 : Registre des aires protégées au Québec}
\source{
https://www.donneesquebec.ca/recherche/dataset/aires-protegees-au-quebec
}
\usage{
get_data0038()
}
\description{
Ces thématiques présentent la compilation des aires protégées au Québec.
}
\details{
This function loads and formats the data
}
\keyword{}
