% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0040-mpa.R
\name{get_data0040}
\alias{get_data0040}
\title{Data 0040 : Oceans Act Marine Protected Areas}
\source{
https://open.canada.ca/data/en/dataset/a1e18963-25dd-4219-a33f-1a38c4971250
}
\usage{
get_data0040()
}
\description{
Fisheries and Oceans Canada has a number of Marine Protected Areas designated under the Oceans Act and Areas of Interest for new Marine Protected Areas at various stages of progress towards designation. These areas are ecologically significant, with species and/or properties that require special management consideration.
}
\details{
This function loads and formats the data
}
\keyword{aires}
\keyword{protégées}
