% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0041-port.R
\name{get_data0041}
\alias{get_data0041}
\title{Data 0041 : Installations portuaires}
\source{
https://ouvert.canada.ca/data/fr/dataset/dcf460ae-c27f-4721-bec4-6e4467a5a3b6
}
\usage{
get_data0041()
}
\description{
Localisation des terminaux portuaires
}
\details{
This function loads and formats the data
}
\keyword{ports}
