% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0043-colonie_oiseaux_cdqs.R
\name{get_data0043}
\alias{get_data0043}
\title{Data 0043 : Computerized Database of Québec Seabirds (CDQS)}
\source{

}
\usage{
get_data0043()
}
\description{
The Computerized Database of Quebec Seabirds (CDQS) was set and has been updated by the Canadian Wildlife Service (CWS), (Quebec region,) for over 20 years. The CDSQ contains information on seabird colonies part of the following species: Gaviidae (loons), Hydrobatidae (petrels), Sulidae (gannets/boobies), Phalacrocoracidae (cormorants), Laridae (seagulls, gulls, terns), Auks (guillemots, penguins, puffins) aud Anatidae.
}
\details{
This function loads and formats the data
}
\keyword{colonies}
\keyword{composantes}
\keyword{d'oiseaux}
\keyword{habitats}
\keyword{valorisées}
