% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0044-sites_patrimoniaux.R
\name{get_data0044}
\alias{get_data0044}
\title{Data 0044 : Sites patrimoniaux}
\source{
https://www.donneesquebec.ca/recherche/fr/dataset/578d182d-4897-4c11-aa7f-ee0cfed222a0
}
\usage{
get_data0044()
}
\description{
Sites patrimoniaux classés par le ministre de la Culture et des Communications
}
\details{
This function loads and formats the data
}
\keyword{composantes}
\keyword{patrimoniaux}
\keyword{sites}
\keyword{valorisées}
