% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0045-sites_patrimoniaux.R
\name{get_data0045}
\alias{get_data0045}
\title{Data 0045 : Sites patrimoniaux}
\source{
https://www.donneesquebec.ca/recherche/fr/dataset/34ca1a49-a020-4f60-986b-69fb4cfa852a
}
\usage{
get_data0045()
}
\description{
Sites patrimoniaux cités par les municipalités et les communautés autochtones
}
\details{
This function loads and formats the data
}
\keyword{composantes}
\keyword{patrimoniaux}
\keyword{sites}
\keyword{valorisées}
