% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0046-dragage_eee.R
\name{get_data0046}
\alias{get_data0046}
\title{Data 0046 : Dragage}
\source{

}
\usage{
get_data0046()
}
\description{
Dragage d'entretien annuel de la voie navigable du Saint-Laurent
}
\details{
Cette fonction formatte les données
}
\keyword{dragage}
\keyword{stresseurs}
