% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0047-zone_portuaire.R
\name{get_data0047}
\alias{get_data0047}
\title{Data 0047 : Délimitation des zones industrialo-portuaires}
\source{
https://www.donneesquebec.ca/recherche/dataset/delimitation-des-zones-ip-industrialo-portuaires
}
\usage{
get_data0047()
}
\description{
Données géolocalisées et cartes illustrant la délimitation des zones industrialo-portuaires
}
\details{
This function loads and formats the data
}
\keyword{ports}
