% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0049-port_contrecoeur.R
\name{get_data0049}
\alias{get_data0049}
\title{Data 0049 : Dragage prévu Port de Montréal à Contrecoeur}
\source{
SNC Lavalin (2017) Agrandissement du terminal portuaire de Contrecoeur. Étude d’impact environnemental - Volume 1 – Rapport principal. Rapport final - F00. Ref. Interne 639223. 889p

SNC Lavalin (2017) Contrecoeur Maritime Terminal Expansion Project. Environmental Impact Assessment – Summary. Final Report > Rev. F02. Internal ref. 639223. 109p.

Impact Assessment Agency of Canada (2021) Contrecoeur Port Terminal Expansion Project. Environmental Assessment Report. Catalog Number: En106-236/2021E-PD. ISBN: 978-0-660-37365-2. 336p. + xiii.
}
\usage{
get_data0049()
}
\description{
Identification des activités de dragage prévus au Port de Montréal à Contrecoeur
}
\details{
Cette fonction formatte les données
}
\keyword{dragage}
\keyword{stresseurs}
