% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0050-port_trois_rivieres.R
\name{get_data0050}
\alias{get_data0050}
\title{Data 0050 : Dragage prévu Port de Trois-Rivières}
\source{
Stantec (2015) Agrandissement des installations portuaires du Port de Trois-Rivières. Description de projet désigné en vertu de la Loi canadienne sur l’évaluation environnementale (2012). Rapport Principal. Administration portuaire de Trois-Rivières. 131-P-0007772-0-01-100-EN-R-0002-01. 91p.
}
\usage{
get_data0050()
}
\description{
Identification des activités de dragage prévus au Port de Trois-Rivières
}
\details{
Cette fonction formatte les données
}
\keyword{dragage}
\keyword{stresseurs}
