% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0051-ballast_exchange.R
\name{get_data0051}
\alias{get_data0051}
\title{Data 0051 : Canada's designated alternate ballast water exchange areas}
\source{
https://open.canada.ca/data/en/dataset/23d26c61-b119-42c0-aa41-bd06cd96a973
}
\usage{
get_data0051()
}
\description{
This data is intended to identify Canadian Alternate Exchange Areas described in https://tc.canada.ca/en/marine-transportation/marine-safety/list-canada-s-designated-alternate-ballast-water-exchange-area-fresh-waters-tp-13617e-2021.
}
\details{
This function loads and formats the data
}
\keyword{area}
\keyword{ballast}
\keyword{exchange}
\keyword{water}
