% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0052-port_quebec.R
\name{get_data0052}
\alias{get_data0052}
\title{Data 0052 : Dragage d'entretien Port de Québec}
\source{
Communication personnelle Marie-Ève Lemieux (marie-eve.lemieux@portquebec.ca)
}
\usage{
get_data0052()
}
\description{
Activités de dragage d'entretien au Port de Québec
}
\details{
Cette fonction formatte les données
}
\keyword{dragage}
\keyword{stresseurs}
