% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0053-milieu_humide.R
\name{get_data0053}
\alias{get_data0053}
\title{Data 0053 : Milieux humides potentiels}
\source{
https://www.donneesquebec.ca/recherche/dataset/milieux-humides-potentiels
}
\usage{
get_data0053()
}
\description{
La cartographie des milieux humides potentiels du Québec (CMHPQ) 2019, diffusée par la Direction de la connaissance écologiques (DCE), fournit une information la plus à jour sur la présence potentielle de milieux humides pour toute la province du Québec. Une série de nouveautés sont maintenant intégrées dans la version 2019 de la CMHPQ. Entre autre, l’ajout de nouvelles sources de données, la classification des entités géographiques potentiellement humide selon une typologie retenue et l’attribution d’un niveau de confiance associé à chaque entité. Cette cartographie constitue une agrégation de différentes bases de données produites à d’autres fins et à des échelles différentes, le résultat de cet assemblage dépend de la précision et de l’exactitude de chacune des sources de données utilisées. Ces données doivent être utilisées en restant vigilant au regard de leurs potentiels et limites.
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{humide}
\keyword{milieu}
\keyword{valorisée}
