% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0054-mammiferes_marins_romm.R
\name{get_data0054}
\alias{get_data0054}
\title{Data 0054 : Observation mammifères marins}
\source{
https://ogsl.ca/fr/biodiversite-mammiferes-marins-romm-a-propos/
}
\usage{
get_data0054()
}
\description{
Data 0054 : Observation mammifères marins
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{mammifères}
\keyword{marins}
\keyword{valorisée}
