% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0055-wendake.R
\name{get_data0055}
\alias{get_data0055}
\title{Data 0055 : Nation Huronne-Wendat - Sites d'intérêt}
\source{
Nation Huronne-Wendat (2021) Localisation des sites Sites d’intérêt culturels, patrimoniaux et archéologiques pour la Nation Huronne-Wendat
}
\usage{
get_data0055()
}
\description{
Localisation des sites Sites d’intérêt culturels, patrimoniaux et archéologiques pour la Nation Huronne-Wendat
}
\details{
Cette fonction formatte les données
}
\keyword{Huronne-Wendat}
\keyword{Nation}
\keyword{Wendake}
\keyword{composante}
\keyword{d'importance}
\keyword{site}
\keyword{valorisée}
