% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0056-gisement_petoncle.R
\name{get_data0056}
\alias{get_data0056}
\title{Data 0056 : Gisements connus et exploités de pétoncles dans l'estuaire et le golfe du Saint-Laurent}
\source{
https://ouvert.canada.ca/data/fr/dataset/82e082be-8f04-4273-830a-8aea63319f07
}
\usage{
get_data0056()
}
\description{
Le pétoncle d'Islande (Chlamys islandica) et le pétoncle géant (Placopecten magellanicus) représentent une ressource biologique importante pour les pêcheries commerciales et l'écologie côtière. Les gisements de cette couche sont décrits comme étant connus et commercialement exploités historiquement et/ou actuellement et ont subi un traitement de validation et de délimitation avec un bon degré de précision. Ils sont situés dans l'estuaire et le golfe du Saint-Laurent, sur la Côte-Nord, Anticosti, la rive Nord de la Gaspésie et les Îles-de-la-Madeleine. Les gisements de cette couche sont considérés parmi les gisements les plus abondants en pétoncles et utilisés pour la pêche commerciale (à noter: les fortes densités et la pêche ne s'en tiennent pas qu'à ces gisements). Ils sont reconnus vulnérables principalement durant la période de croissance larvaire d'août à novembre pour la survie des larves, mais sont vulnérables aussi le reste de l'année étant des espèces sessiles. La couche de données contient les informations sur l'aire totale des gisements, la zone et le fond de pêche (étape de validation), le nom des gisements, les périodes de reproduction par secteur et les périodes de pêche (valide en 2015).
}
\details{
This function loads and formats the data
}
\keyword{pétoncle}
