% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0057-gisement_mactre.R
\name{get_data0057}
\alias{get_data0057}
\title{Data 0057 : Gisements connus et exploités de mactre de Stimpson des eaux côtières du Québec}
\source{
https://ouvert.canada.ca/data/fr/dataset/8bb29ee0-6cd8-4dd4-abe0-afe8682a69d9
}
\usage{
get_data0057()
}
\description{
La mactre de Stimpson est visée par des mesures de protection et de conservation ayant pour objectif de protéger son potentiel reproducteur et l'intégrité des population (gisements) dans la région du Québec. Les gisements de cette couche sont décrits comme étant connus et commercialement exploités historiquement et/ou actuellement et ont subi un traitement de validation et de délimitation avec un bon degré de précision. Ils sont situés dans l'estuaire et le Golfe du Saint-Laurent, sur la Côte-Nord et les Îles-de-la-Madeleine. Les gisements de cette couche sont considérés parmi les gisements les plus abondants en mactre de Stimpson et sont utilisés pour la pêche commerciale (à noter: les fortes densités ne s'en tiennent pas qu'à ces gisements). Ils sont principalement reconnus vulnérables pendant la saison de reproduction de juin à juillet. Il pourrait y avoir une ponte secondaire à l'automne également. La couche de données contient les informations sur l'aire totale des gisements, les zones et noms des gisements, les secteurs concernés et les périodes de reproduction et de pêcheries.
}
\details{
This function loads and formats the data
}
\keyword{Stimpson}
\keyword{de}
\keyword{mactre}
