% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0058-zico.R
\name{get_data0058}
\alias{get_data0058}
\title{Data 0058 :}
\source{
https://www.ibacanada.com/explore_how.jsp?lang=fr
}
\usage{
get_data0058()
}
\description{
Les zones importantes pour la conservation des oiseaux (ZICO) sont des sites qui abritent des groupes d’oiseaux bien précis : espèces menacées, vaste population d’oiseaux ou espèces ayant une aire de répartition réduite. De tailles variées, les ZICO peuvent couvrir de minuscules parcelles de territoire comme de vastes étendues de terre, des rivières et des plans d’eau. Elles peuvent être constituées de terres publiques ou privées et peuvent chevaucher en tout ou en partie des aires protégées.
}
\details{
This function loads and formats the data
}
\keyword{oiseaux}
\keyword{zico}
