% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0059-cdpnq_faune.R
\name{get_data0059}
\alias{get_data0059}
\title{Data 0059 : Espèces fauniques à statut}
\source{

}
\usage{
get_data0059()
}
\description{
Présence d'espèces fauniques menacées ou vulnérables ou susceptibles d'être ainsi désignées ou rares
}
\details{
Cette fonction formatte les données
}
\keyword{composante}
\keyword{espèce}
\keyword{faune}
\keyword{habitat}
\keyword{statut}
\keyword{valorisée}
\keyword{à}
