% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0060-cdpnq_flore.R
\name{get_data0060}
\alias{get_data0060}
\title{Data 0060 : Espèces floristiques à statut}
\source{

}
\usage{
get_data0060()
}
\description{
Présence d'espèces floristiques menacées ou vulnérables ou susceptibles d'être ainsi désignées ou rares
}
\details{
Cette fonction formatte les données
}
\keyword{composante}
\keyword{espèce}
\keyword{flore}
\keyword{habitat}
\keyword{statut}
\keyword{valorisée}
\keyword{à}
