% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0061-vulnerability_habitat_kappel.R
\name{get_data0061}
\alias{get_data0061}
\title{Data 0061 : Vulnerability scores for 58 stressors in 15 ecosystems}
\source{
Kappel, Halpern, and Napoli. (2012, January) Mapping Cumulative Impacts of Human Activities on Marine Ecosystems (03.NCEAS.12) Boston: SeaPlan.
}
\usage{
get_data0061()
}
\description{
Matrix of vulnerability scores for 58 different types of stressors in 15 types of intertidal, subtidal and offshore ecosystems by Kappel et al. (2012).
}
\details{
Cette fonction formatte les données
}
\keyword{habitat}
\keyword{vulnérabilité}
