% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0062-vessel_of_concern_inventory.R
\name{get_data0062}
\alias{get_data0062}
\title{Data 0062 : Inventory of vessel of concern}
\source{
Canadian Coast Guard
}
\usage{
get_data0062()
}
\description{
Distribution of vessels of concern filtered for our study area. The majority of vessels have yet to be validated by CCG: this validation process is underway. As such, the accuracy of the coordinates cannot be guaranteed.
}
\details{
Cette fonction formatte les données
}
\keyword{naufrages}
\keyword{stresseurs}
