% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0064-vulnerability_mammiferes_marins.R
\name{get_data0064}
\alias{get_data0064}
\title{Data 0064 : Scores de vulnérabilité pour les mammifères marins}
\source{
Our team
}
\usage{
get_data0064()
}
\description{
Data 0064 : Scores de vulnérabilité pour les mammifères marins
}
\details{
Cette fonction formatte les données
}
\keyword{mammifères}
\keyword{marins}
\keyword{vulnérabilité}
