% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0065-decoupages_administratifs.R
\name{get_data0065}
\alias{get_data0065}
\title{Data 0065 : Découpages administratifs}
\source{
https://www.donneesquebec.ca/recherche/dataset/decoupages-administratifs
}
\usage{
get_data0065()
}
\description{
Les couches de découpages administratifs offertes aux échelles de 1/20 000 et 1/100 000 permettent la localisation des limites pour les composantes suivantes : 1) Les arrondissements et agglomérations; 2) Les municipalités, territoires non organisés et territoires autochtones; 3)Les régions administratives, communautés métropolitaines, et MRC; 4)La frontière internationale, les frontières interprovinciales ainsi que la frontière Québec–Terre-Neuve-et-Labrador.
}
\details{
This function loads and formats the data
}
\keyword{administratifs}
\keyword{découpages}
