% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0067-gcnwa_site_archeologique.R
\name{get_data0067}
\alias{get_data0067}
\title{Data 0067 : Sites archéologiques GCNWA}
\source{
Grand Conseil de la Nation Waban-Aki (GCNWA), 2021. Rapport d’intrants fournis par le Bureau du Ndakina quant à l’étude des effets cumulatifs de la navigation commerciale. Rapport à l’intention de Transport Canada et Pêches et Océans Canada. Rédigé par E. Blanchet et G. Treyvaud, Bureau du Ndakina, 42 p. et annexes.
}
\usage{
get_data0067()
}
\description{
Rapport d’intrants fournis par le Bureau du Ndakina quant à l’étude des effets cumulatifs de la navigation commerciale. Rapport à l’intention de Transport Canada et Pêches et Océans Canada.
}
\details{
Cette fonction formatte les données
}
\keyword{GCNWA}
\keyword{Sites}
\keyword{archéologiques}
\keyword{composante}
\keyword{valorisée}
