% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0068-gcnwa_site_potentiel_archeologique.R
\name{get_data0068}
\alias{get_data0068}
\title{Data 0068 : Sites à potentiel archéologique GCNWA}
\source{
Grand Conseil de la Nation Waban-Aki (GCNWA), 2021. Rapport d’intrants fournis par le Bureau du Ndakina quant à l’étude des effets cumulatifs de la navigation commerciale. Rapport à l’intention de Transport Canada et Pêches et Océans Canada. Rédigé par E. Blanchet et G. Treyvaud, Bureau du Ndakina, 42 p. et annexes.
}
\usage{
get_data0068()
}
\description{
Rapport d’intrants fournis par le Bureau du Ndakina quant à l’étude des effets cumulatifs de la navigation commerciale. Rapport à l’intention de Transport Canada et Pêches et Océans Canada.
}
\details{
Cette fonction formatte les données
}
\keyword{GCNWA}
\keyword{Site}
\keyword{archéologique}
\keyword{composante}
\keyword{potentiel}
\keyword{valorisée}
\keyword{à}
