% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0069-dragage_essipit.R
\name{get_data0069}
\alias{get_data0069}
\title{Data 0069 : Sites de dragage et de dépôt de la Première Nation des Innus d'Essipit}
\source{
Bourchard D. (communication personnelle) Position et volume dragués et den 2001, 2011 et 2016 pour la Première Nation des Innus d'Essipit dans la région des Bergeronnes
}
\usage{
get_data0069()
}
\description{
Position et volume des sites de dragages et de dépôt en 2001, 2011 et 2016 pour la Première Nation des Innus d'Essipit dans la région des Bergeronnes
}
\details{
Cette fonction formatte les données
}
\keyword{dragage}
