% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0070-ancrage_gcc.R
\name{get_data0070}
\alias{get_data0070}
\title{Data 0070 : Sites d'ancrage commerciaux}
\source{
Garde côtière canadienne (2022) Fiches techniques de mouillage - Les Escoumins à Montréal.
}
\usage{
get_data0070()
}
\description{
Localisation des sites d'ancrage commerciaux provenant des fiches techniques de mouillage entre Les Escoumins et Montréal
}
\details{
Cette fonction formatte les données
}
\keyword{ancrage}
