% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0071-ancrage_shc.R
\name{get_data0071}
\alias{get_data0071}
\title{Data 0071 : Mouillages et zones de mouillages canadiennes}
\source{

}
\usage{
get_data0071()
}
\description{
Mouillages et zones de mouillages canadiennes est composés de quatre fichiers qui comprennent des zones de mouillage (ACHARE) et des mouillages à un seul navire (ACHBRT). Les fichiers sont soit un avec des points et un avec des polygones pour chacun des deux types mouillages. Ces fichiers sont un ensemble complet de données sur les mouillages et les zones de mouillage dans les eaux navigables canadiennes. Cet ensemble de données sera mis à jour au besoin.
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{ancrage}
\keyword{stresseurs}
