% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/data0072-wolastoqiyik_wahsipekuk_sites_interet.R
\name{get_data0072}
\alias{get_data0072}
\title{Data 0072 : Communauté Wolastoqiyik Wahsipekuk sites d'intérêt}
\source{
Première Nation Wolastoqiyik Wahsipekuk (2022) Localisation de sites d'intérêt pour l'observation du béluga et le développement portuaire
}
\usage{
get_data0072()
}
\description{
Localisation de sites d'intérêt pour l'observation du béluga et le développement portuaire
}
\details{
Cette fonction formatte les données
}
\keyword{Wahsipekuk}
\keyword{Wolastoqiyik}
\keyword{composante}
\keyword{d'interêt}
\keyword{sites}
\keyword{valorisée}
