% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0073-reproduction_alose_savoureuse.R
\name{get_data0073}
\alias{get_data0073}
\title{Data 0073 : Aires de reproduction de l'alose savoureuse dans le fleuve et l'estuaire du Saint-Laurent}
\source{
https://ouvert.canada.ca/data/fr/dataset/a0817db6-42cb-4fa2-a601-48246c812e94
}
\usage{
get_data0073()
}
\description{
Couche regroupant l’information connue des aires de reproduction de l'alose savoureuse dans le fleuve et l'estuaire du Saint-Laurent selon une revue de littérature de documents réalisés entre 1976 et 1997.
}
\details{
This function loads and formats the data
}
\keyword{frayères}
