% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0074-reproduction_esturgeon_jaune.R
\name{get_data0074}
\alias{get_data0074}
\title{Data 0074 : Aires de reproduction, d'alimentation et de concentration de l'esturgeon jaune dans le fleuve et l'estuaire du Saint-Laurent}
\source{
https://ouvert.canada.ca/data/fr/dataset/0db67020-7521-4356-bd53-3132cb7804a3
}
\usage{
get_data0074()
}
\description{
Couche regroupant l’information connue des aires de reproduction, d'alimentation et de concentration de l'esturgeon jaune dans le fleuve et l'estuaire du Saint-Laurent selon une revue de littérature de documents réalisés entre 1976 et 2002.
}
\details{
This function loads and formats the data
}
\keyword{frayères}
