% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0075-reproduction_esturgeon_noir.R
\name{get_data0075}
\alias{get_data0075}
\title{Data 0075 : Aires de reproduction, d'alimentation et de concentration de l'esturgeon noir dans le fleuve et l'estuaire du Saint-Laurent}
\source{
https://ouvert.canada.ca/data/fr/dataset/bfa19b1c-a248-4dc6-b497-327a8937a07b
}
\usage{
get_data0075()
}
\description{
Couche regroupant l’information connue des aires de reproduction, d'alimentation et de concentration de l'esturgeon noir dans le fleuve et l'estuaire du Saint-Laurent selon une revue de littérature de documents réalisés entre 1993 et 2003.
}
\details{
This function loads and formats the data
}
\keyword{frayères}
