% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0076-reproduction_eperlan_arc_en_ciel.R
\name{get_data0076}
\alias{get_data0076}
\title{Data 0076 : Aires de reproduction et d'alimentation de l'éperlan-arc-en-ciel dans la Rivière Saguenay, le fleuve et l'estuaire du Saint-Laurent}
\source{
https://ouvert.canada.ca/data/fr/dataset/9ddbf6ff-4314-47d8-ac0f-8d2f7278edcf
}
\usage{
get_data0076()
}
\description{
Couche regroupant l’information connue des aires de reproduction et d'alimentation de l'éperlan-arc-en-ciel, dans la Rivière Saguenay, le fleuve et l'estuaire du Saint-Laurent selon une revue de littérature de documents réalisés entre 1977 et 2000.
}
\details{
This function loads and formats the data
}
\keyword{frayères}
