% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0077-habitat_essentiel_LEP.R
\name{get_data0077}
\alias{get_data0077}
\title{Data 0077 : L'habitat essentiel désigné en vertu de la Loi sur les espèces en péril}
\source{
https://ouvert.canada.ca/data/fr/dataset/db177a8c-5d7d-49eb-8290-31e6a45d786c
}
\usage{
get_data0077()
}
\description{
Le Programme des espèces en péril consiste à réaliser le mandat du MPO en vertu de la Loi sur les espèces en péril (LEP) afin d’assurer la protection, le rétablissement et la conservation de toutes les espèces aquatiques en péril inscrites au Canada.
}
\details{
This function loads and formats the data
}
\keyword{frayères}
