% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0078-clovisse_arctique.R
\name{get_data0078}
\alias{get_data0078}
\title{Data 0078 : Aires de concentration connues de la Clovisse arctique dans l’estuaire et le golfe du Saint-Laurent}
\source{
https://open.canada.ca/data/fr/dataset/b0633101-18e6-483a-b668-4e1f76789696
}
\usage{
get_data0078()
}
\description{
La couche représente des aires de concentration connues de la Clovisse arctique (Mesodesma arctatum) exploitées (communément appelées 'gisements') ou non exploitées de l'estuaire et du golfe du Saint-Laurent, région du Québec. Elle a été conçue pour le Centre national des urgences environnementales (CNUE) pour la préparation et l'intervention en cas de déversements pétroliers. Les aires de concentration ont été délimitées grâce à des inventaires réalisés par Pêches et Océans Canada (MPO) entre les années 2000 et 2020 et des données de divers projets de recherche du MPO. Pour plus d’information sur la façon dont la couche de données a été construite, consultez les métadonnées incluses dans son fichier de forme (.shp), particulièrement la section 'Lineage'.
}
\details{
This function loads and formats the data
}
\keyword{coquilliers}
\keyword{gisements}
