% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0079-couteau_atlantique.R
\name{get_data0079}
\alias{get_data0079}
\title{Data 0079 : Aires de concentration connues du Couteau de l'Atlantique dans l’estuaire et le golfe du Saint-Laurent}
\source{
https://open.canada.ca/data/fr/dataset/a0039b2e-7847-4d48-8fd8-638677760f8b
}
\usage{
get_data0079()
}
\description{
La couche représente des aires de concentration du Couteau de l'Atlantique (Ensis leei) exploitées (communément appelées « gisements ») ou non exploitées de l’estuaire et du golfe du Saint-Laurent dans la région du Québec. Elle a été conçue pour le Centre national des urgences environnementales (CNUE) pour la préparation et l’intervention en cas de déversements d’hydrocarbures. Les aires de concentration ont été délimitées grâce à des inventaires et divers projets de recherche réalisés par Pêches et Océans Canada (MPO) et de données de pêches commerciales.
}
\details{
This function loads and formats the data
}
\keyword{coquilliers}
\keyword{gisements}
