% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0080-mactre_atlantique.R
\name{get_data0080}
\alias{get_data0080}
\title{Data 0080 : Aires de concentration connues de la Mactre de l'Atlantique dans le golfe du Saint-Laurent}
\source{
https://open.canada.ca/data/fr/dataset/ef06ae4e-151c-439c-be88-ef3b6d3b39a6
}
\usage{
get_data0080()
}
\description{
La couche représente des aires de concentration connues de la Mactre de l'Atlantique (Spisula solidissima) exploitées (communément appelées « gisements ») ou non exploitées présentes dans le golfe du Saint-Laurent, autour des Îles-de-la-Madeleine. Elle a été conçue pour le Centre national des urgences environnementales (CNUE) pour la préparation et l’intervention en cas de déversements d’hydrocarbures. Les aires de concentration ont été délimitées grâce à des données d'inventaire de Pêches et Océans Canada (MPO) de 2000 à 2020, des données de pêches commerciales (journaux de bord et récépissés d'achat de 2002 à 2020 et observations en mer de 2005 à 2017), des données de pêches exploratoires ainsi que des entrevues.
}
\details{
This function loads and formats the data
}
\keyword{coquilliers}
\keyword{gisements}
