% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0081-mye_commune.R
\name{get_data0081}
\alias{get_data0081}
\title{Data 0081 : Aires de concentration connues de la Mye commune dans la zone intertidale de l’estuaire et du golfe du Saint-Laurent}
\source{
https://open.canada.ca/data/fr/dataset/83fc6c4f-8a13-4e42-8d22-4dbacb5fa8c3
}
\usage{
get_data0081()
}
\description{
La couche représente des aires de concentration de la Mye commune (Mya arenaria) exploitées (communément appelées « gisements ») ou non exploitées présentes dans la zone intertidale de l’estuaire et du golfe du Saint-Laurent de la région du Québec. Elle a été conçue pour le Centre national des urgences environnementales (CNUE) pour la préparation et l’intervention en cas de déversement pétrolier. Les aires de concentration ont été délimitées grâce à des inventaires réalisés par le Pêches et Océans Canada (MPO) entre les années 2000 et 2020. À noter que cette couche est tributaire des inventaires effectués et ne représente ainsi que les aires de Mye commune connues. Par exemple, pour la Haute-Côte-Nord, les inventaires ont été limités aux secteurs ouverts à la cueillette (à l'exception de 4 secteurs), mais il est connu que la Mye commune est présente aussi à l'extérieur de ces secteurs. De plus, peu d'informations étaient disponibles pour la Moyenne et la Basse-Côte-Nord.
}
\details{
This function loads and formats the data
}
\keyword{coquilliers}
\keyword{gisements}
