% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0082-aires_protegees_publiques_privees.R
\name{get_data0082}
\alias{get_data0082}
\title{Data 0082 : Aires protégées publiques et privées}
\usage{
get_data0082()
}
\description{
Couche de données intégrées qui inclue les aires protégées inscrites au Registre des Aires protégées du Québec, les territoires d’importance, le Réseau de Milieux Naturels protégés (RMN) en plus des habitats fauniques et des écosystèmes forestiers exceptionnels qui ne sont pas inscrits au registre. Les multiparties ont été séparées en partie et les coordonnées des centroides de chaque polygone ont été ajoutées pour servir de localisation gps. Cette couche a été intégrée le 2022/10/13.
}
\details{
Cette fonction formatte les données
}
