% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0083-pressions_saint_laurent.R
\name{get_data0083}
\alias{get_data0083}
\title{Data 0083 : Cartographie des principales pressions exercées sur le Saint-Laurent en 2012}
\usage{
get_data0083()
}
\description{
Cartographie de six pressions susceptibles d’avoir des impacts sur les poissons dans le fleuve Saint-Laurent: agriculture, artificialisation des berges, navigation commerciale, pêche commerciale, régularisation du débit, industrialisation/urbanisation. Un indice a été calculé pour chacune d’elles et aux fins de comparaison, les valeurs ont été remplacées par les cotes 1, 3 ou 5 correspondant aux niveaux faible, moyen ou fort. Pour cartographier les pressions, ces calculs ont été effectués dans chaque segment du fleuve, correspondant à des unités de 5 km de long partant des rives nord ou sud et allant jusqu’à la voie navigable.
}
\details{
Cette fonction formatte les données
}
