% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0084-biovolumes_herbiers.R
\name{get_data0084}
\alias{get_data0084}
\title{Data 0084 : Biovolumes herbiers}
\usage{
get_data0084()
}
\description{
Biovolumes d'herbiers aquatiques à partir de données sonar en 2021.
}
\details{
Cette fonction formatte les données
}
