% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0085-large_whales_dfo.R
\name{get_data0085}
\alias{get_data0085}
\title{Data 0085 : Seasonal distribution and concentration of four baleen whale species in the St. Lawrence Estuary}
\usage{
get_data0085()
}
\description{
The Estuary and the Gulf of St. Lawrence are feeding grounds for several North Atlantic whales, including minke, humpback, endangered blue whales and fin whales which are designated by COSEWIC as Special Concern. Observations of these species collected during aerial and boat surveys conducted between 1995 to 2017 were used to produce maps presenting raw sightings, relative densities from kernel analyses and predicted relative probability of occurrence from spatial distribution modelling.
}
\details{
Cette fonction formatte les données
}
