% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataXXXX-conservation.R
\name{get_data00XXX}
\alias{get_data00XXX}
\title{Data 0033 : Territoires d’intérêt pour la conservation}
\source{
https://www.donneesquebec.ca/recherche/dataset/atlas-des-territoires-interet-conservation-btsl#
}
\usage{
get_data00XXX()
}
\description{
Atlas des territoires d’intérêt pour la conservation dans les Basses-terres du Saint-Laurent
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{conservation}
\keyword{d'impoetance}
\keyword{d'intérêt}
\keyword{la}
\keyword{pour}
\keyword{site}
\keyword{territoire}
\keyword{valorisée}
