% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataXXXX-Qualite_eau.R,
%   R/dataXXXX-depot_surface.R, R/dataXXXX-reseay_hydrographique.R,
%   R/dataXXXX-suivi_bacteriologique.R
\name{get_dataXXXX}
\alias{get_dataXXXX}
\title{Data XXXX : Suivi de la qualité de l'eau du fleuve et des rivières}
\source{
https://www.donneesquebec.ca/recherche/dataset/suivi-physicochimique-des-rivieres-et-du-fleuve#

https://www.donneesquebec.ca/recherche/dataset/depots-de-surface

https://www.donneesquebec.ca/recherche/dataset/grhq

https://www.donneesquebec.ca/recherche/dataset/suivi-bacteriologique-fleuve-sites-potentiels-baignade
}
\usage{
get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()
}
\description{
Data XXXX : Suivi de la qualité de l'eau du fleuve et des rivières

Data XXXX : Dépôts de surface

Data XXXX : Géobase du réseau hydrographique du Québec (GRHQ)

Data XXXX : Suivi bactériologique du fleuve - Sites potentiels de baignade
}
\details{
This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data
}
\keyword{a}
