% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0000_zonesNA.R
\name{get_zonesNA}
\alias{get_zonesNA}
\title{Zones for NA values (this could definitely be improved...)}
\usage{
get_zonesNA()
}
\description{
Zones for NA values (this could definitely be improved...)
}
