% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-global_parameters.R
\name{global_parameters}
\alias{global_parameters}
\title{Loads global parameters}
\usage{
global_parameters()
}
\description{
The function loads the global parameters for this project
}
\examples{
# Metadata
global_parameters()
}
