% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_lissage.R
\name{lissage}
\alias{lissage}
\title{Lissage Gaussien de 0.2o}
\usage{
lissage(dat)
}
\arguments{
\item{dat}{observation points as sf object}
}
\description{
Lissage Gaussien de 0.2o
}
\keyword{gaussien}
\keyword{kernel}
\keyword{lissage}
