% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_load.R
\name{load_format}
\alias{load_format}
\alias{load_metadata}
\alias{load_contact}
\alias{load_integrated}
\alias{basemap}
\title{Load data, metadata and contact information}
\usage{
load_format(data_id)

load_metadata(data_id)

load_contact(data_id)

load_integrated(data_name)

basemap(basemap_name)
}
\arguments{
\item{data_id}{\code{character} id of data to import in R session with format \code{dataXXXX} for \code{load_format()}, \code{metadata_dataXXXX} for \code{load_metadata()}, and \code{contact_dataXXXX} for \code{load_contact()}.}

\item{data_name}{\code{character} name of integrated data to load in R session with format, one of \code{navigation}, \code{peche}, \code{deversement}, \code{dragage}, \code{ancrage}, \code{rejet}, \code{epaves} for stressors, and one of \code{habitat}, \code{site}, \code{berge}, \code{mammifere_marin}, \code{quality} for valued components.}

\item{basemap_name}{\code{character} name of basemap data to load, one of \code{egsl}, \code{quebec}, \code{canada}, \code{usa}}
}
\description{
The function imports data available in \verb{./data/data-format/} regardless of its format.
}
\details{
This function is used to avoid putting all formatted dataset in lazy load, yet still make it easy for them to be loaded. The user simply provides the id of the dataset and the function imports it in the R session, along with its metadata and contact information if desired
}
\examples{
load_format("data0001")
}
\keyword{contact}
\keyword{data}
\keyword{metadata}
