% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_metadata_st_cv.R
\name{metadata_st_cv}
\alias{metadata_st_cv}
\title{Generate metadata for stressors}
\usage{
metadata_st_cv()
}
\description{
Generate metadata for stressors
}
\details{
These functions take '.yml' files in specified folder to generate an individual table containing information from all \code{.yml} files
}
\keyword{metadata}
\keyword{stresseurs}
