% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_observation_distribution.R
\name{observation_distribution}
\alias{observation_distribution}
\title{Function to evaluate the distribution of observations for marine mammals}
\usage{
observation_distribution(dat, studygrid)
}
\arguments{
\item{dat}{observation points for marine mammals as sf object}

\item{studygrid}{regular grid as sf object}
}
\description{
The function evaluates the frequency of marine mammal observations as in data0027
}
\keyword{mammifères}
\keyword{marins}
