% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline d'analyse}
\usage{
pipeline(
  pipeline_metadata = FALSE,
  pipeline_data = FALSE,
  pipeline_integration = FALSE,
  pipeline_analysis = FALSE,
  pipeline_figures = FALSE,
  pipeline_annexes = FALSE,
  pipeline_report = TRUE
)
}
\description{
Pipeline d'analyse du projet d'évaluation des effets cumulatifs de la navigation dans le Saguenay et le Saint-Laurent
}
\details{
Cette fonction exécute l'ensemble des fonctions pour effectuer l'analyse des effets cumulatifs
}
\keyword{pipeline}
