% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_plot_ceanav.R
\name{plot_ceanav}
\alias{plot_ceanav}
\alias{plot_ceanav.sf}
\title{plot ceanav data maps}
\usage{
plot_ceanav(dat, ...)

\method{plot_ceanav}{sf}(
  dat,
  main = NULL,
  subtitle = NULL,
  unit_data = NULL,
  references = NULL,
  city = TRUE,
  ...
)
}
\arguments{
\item{dat}{object of class sf}

\item{...}{further specifications, see \link{plot} and details.}

\item{main}{main title}

\item{subtitle}{subtitle}

\item{unit_data}{units of data}

\item{references}{data citation used for integrated data}

\item{city}{logical, if TRUE name of cities are added to graph}
}
\description{
base plot functions for ceanav project
}
\examples{
 load_integrated("navigation")
 plot_ceanav(navigation[,10], "Navigation", "Recherche gouvernementale")

}
