% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_plot.R
\name{plot_format}
\alias{plot_format}
\alias{plot_format.sf}
\alias{plot_format.stars}
\title{plot ceanav formatted data}
\usage{
plot_format(dat, ...)

\method{plot_format}{sf}(dat, ...)

\method{plot_format}{stars}(dat, ...)
}
\arguments{
\item{dat}{object of class sf, stars}

\item{...}{further specifications, see \link{plot} and details.}
}
\description{
base plot functions for ceanav project
}
