% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_plot_integrated.R
\name{plot_integrated}
\alias{plot_integrated}
\alias{plot_integrated.sf}
\title{plot ceanav integrated data}
\usage{
plot_integrated(dat, ...)

\method{plot_integrated}{sf}(dat, main = NULL, subtitle = NULL, ...)
}
\arguments{
\item{dat}{object of class sf}

\item{...}{further specifications, see \link{plot} and details.}

\item{main}{main title}

\item{subtitle}{subtitle}
}
\description{
base plot functions for ceanav project
}
\examples{
 load_integrated("navigation")
 plot_integrated(navigation[,10], "Navigation", "Recherche gouvernementale")

}
