% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_plot_simple.R
\name{plot_simple}
\alias{plot_simple}
\alias{plot_simple.sf}
\title{plot ceanav simple data maps with only outline of study area}
\usage{
plot_simple(dat, ...)

\method{plot_simple}{sf}(dat, ...)
}
\arguments{
\item{dat}{object of class sf}

\item{...}{further specifications, see \link{plot} and details.}
}
\description{
base plot functions for ceanav project
}
\examples{
 load_integrated("navigation")
 plot_simple(navigation[,10])

}
