% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_annexe_data_summary.R
\name{rep_annexe_data_summary}
\alias{rep_annexe_data_summary}
\title{Tableau récapitulatif des données}
\usage{
rep_annexe_data_summary()
}
\description{
Fonction utilisée pour générer un tableau récapitulatif des données présentées dans le rapport.
}
\details{
Cette fonction permet de générer le tableau récapitulatif des données utilisées pour le rapport
}
\examples{
rep_annexe_data_summary()
}
\keyword{récapitulatif}
\keyword{tableau}
