% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_contact_summary.R
\name{rep_contact_summary}
\alias{rep_contact_summary}
\title{Tableau récapitulatif des personnes et organisations ressources}
\usage{
rep_contact_summary()
}
\description{
Fonction utilisée pour générer un tableau récapitulatif des personnes et organisations ressources pour le projet.
}
\details{
Cette fonction permet de générer le tableau récapitulatif des personnes et organisations ressources pour le projet
}
\examples{
rep_contact_summary()
}
\keyword{récapitulatif}
\keyword{tableau}
