% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_data_description.R
\name{rep_data_description}
\alias{rep_data_description}
\alias{rep_data_description_all}
\title{Fiches descriptives des données}
\usage{
rep_data_description(data_id, output_folder, suffix = NULL)

rep_data_description_all(output_folder, suffix)
}
\arguments{
\item{data_id}{\code{character} id of data to import in R session with format \code{dataXXXX}}

\item{output_folder}{\code{character} output folder for exported document}

\item{suffix}{\code{character} if provided, add suffix before data name}
}
\description{
Fonction utilisée pour générer une fiches descriptive pour chaque données incluse au tableau récapitulatif des données à l'Annexe 1
}
\details{
Cette fonction permet de générer une fiche descriptive pour la base de données sélectionnée
}
\examples{
rep_data_description("data0001","report/contenu/annexes/","annexe2-")
}
\keyword{descriptive}
\keyword{fiche}
