% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_data_figures.R
\name{rep_data_figures}
\alias{rep_data_figures}
\alias{rep_data_figures_all}
\title{Figures données formatées}
\usage{
rep_data_figures(data_id)

rep_data_figures_all()
}
\description{
Fonctions pour générer des figures pour les données formatées
}
\examples{
rep_data_figures(data_id = "data0001")
}
\keyword{figure}
