% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_data_figures.R, R/rep_data_summary.R
\name{rep_data_summary}
\alias{rep_data_summary}
\title{Figures données formatées}
\usage{
rep_data_summary()

rep_data_summary()
}
\description{
Fonctions pour générer des figures pour les données formatées

Fonction utilisée pour générer un tableau récapitulatif des données présentées dans le rapport.
}
\details{
Cette fonction permet de générer le tableau récapitulatif des données utilisées pour le rapport
}
\examples{
rep_data_figures(data_id = "data0001")
rep_data_summary()
}
\keyword{figure}
\keyword{récapitulatif}
\keyword{tableau}
