% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_hyperlien.R
\name{rep_hyperlien}
\alias{rep_hyperlien}
\title{Hyperlien à partir de deux vecteurs}
\usage{
rep_hyperlien(texte, url)
}
\arguments{
\item{texte}{\code{character}, texte à utiliser pour créer l'hyperlien}

\item{url}{\code{character}, hyperlien vers la ressource externe}
}
\description{
Création d'un hyperlien en format markdown à partir de deux vecteurs
}
\details{
Cette fonction permet de générer le tableau récapitulatif des données utilisées pour le rapport
}
\examples{
texte <- "Rapport"
url <- "https://github.com/EffetsCumulatifsNavigation/ceanav"
rep_hyperlien(texte, url)
}
\keyword{hyperlien}
