% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_portrait_data_description.R
\name{rep_portrait_data_description}
\alias{rep_portrait_data_description}
\alias{rep_portrait_data_description_all}
\title{Fiches descriptives des données intégrés}
\usage{
rep_portrait_data_description(data_id, output_folder, suffix)

rep_portrait_data_description_all(output_folder, suffix)
}
\arguments{
\item{data_id}{\code{character} id of data to import in R session}

\item{output_folder}{\code{character} output folder for exported document}

\item{suffix}{\code{character} if provided, add suffix before data name}
}
\description{
Fonction utilisée pour générer une fiches descriptive pour chaque catégorie des données intégrées pour les stresseurs environnementaux et les composantes valorisées
}
\details{
Cette fonction permet de générer une fiche descriptive pour la base de données sélectionnée
}
\examples{
rep_portrait_data_description("int_st_peche_commerciale","report/contenu/5-portrait/1-stresseurs/","peche_commerciale")
}
\keyword{descriptive}
\keyword{fiche}
