% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_deversement.R
\name{st_deversement}
\alias{st_deversement}
\title{Déversements accidentels}
\usage{
st_deversement()
}
\description{
Couche de données transformées pour les déversements accidentels dans le Saint-Laurent
}
\details{
Cette fonction importe et formatte les données pour l'analyse d'effets cumulatifs
}
\keyword{accidentel}
\keyword{déversement}
\keyword{stresseurs}
