% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_dragage.R
\name{st_dragage}
\alias{st_dragage}
\title{Dragage}
\usage{
st_dragage()
}
\description{
Couche de données transformées pour les activités de dragage dans le Saint-Laurent
}
\details{
Cette fonction importe et formatte les données pour l'analyse d'effets cumulatifs
}
\keyword{dragage}
\keyword{stresseurs}
