% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_naufrage.R
\name{st_naufrage}
\alias{st_naufrage}
\title{Naufrages}
\usage{
st_naufrage()
}
\description{
Couche de données présentant les sites de naufrages dans la zone d'étude
}
\details{
Cette fonction importe et formatte les données pour l'analyse d'effets cumulatifs
}
\keyword{naufrage}
\keyword{stresseurs}
