% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_navigation.R
\name{st_navigation}
\alias{st_navigation}
\title{Navigation}
\usage{
st_navigation()
}
\description{
Couche de données transformées pour la navigation dans le Saint-Laurent
}
\details{
Cette fonction importe et formatte les données pour l'analyse d'effets cumulatifs
}
\keyword{navigation}
\keyword{stresseurs}
