% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_pollution_maritime.R
\name{st_pollution_maritime}
\alias{st_pollution_maritime}
\title{Pollution maritime}
\usage{
st_pollution_maritime()
}
\description{
Couche de données transformées pour la pollution maritime dans le Saint-Laurent
}
\details{
Cette fonction importe et formatte les données pour l'analyse d'effets cumulatifs
}
\keyword{ancrage}
\keyword{stresseurs}
