% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_port.R
\name{st_port}
\alias{st_port}
\title{Influence de la navigation en zone portuaire}
\usage{
st_port()
}
\description{
Couche de données transformées pour l'influence de la navigation en zone portuaire dans le Saint-Laurent et le Saguenay
}
\details{
Cette fonction importe et formatte les données pour l'analyse d'effets cumulatifs
}
\keyword{navigation}
\keyword{portuaire}
\keyword{stresseurs}
