% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_metadata.R
\name{update_metadata}
\alias{update_metadata}
\alias{update_contact}
\title{Update metadata and contacts}
\usage{
update_metadata(type)

update_contact()
}
\arguments{
\item{type}{type of metadata, one of \code{raw} or \code{integrated}}
}
\description{
The function updates the existing metadata file based on individual metadata files in \code{data-metadata}
}
\details{
These functions take '.yml' files in specified folder to generate an individual table containing information from all \code{.yml} files
}
\examples{
# Metadata
update_metadata("raw")

# Contacts
update_contact()
}
\keyword{contact}
\keyword{metadata}
