% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuln_berge.R
\name{vuln_berge}
\alias{vuln_berge}
\title{Vulnérabilité des berges à l'érosion causée par les stresseurs environnementaux}
\usage{
vuln_berge()
}
\description{
Création de la matrice de vulnérabilité des berges à l'érosion pour l'évaluation des effets cumulatifs
}
\details{
Cette fonction génère la matrice de vulnérabilité pour les habitats
}
\keyword{berges}
\keyword{composante}
\keyword{des}
\keyword{intégrité}
\keyword{valorisée}
\keyword{vulnérabilité}
