% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuln_faune_flore.R
\name{vuln_faune_flore}
\alias{vuln_faune_flore}
\title{Vulnérabilité des sites d'importance fauniques et floristiques aux stresseurs environnementaux}
\usage{
vuln_faune_flore()
}
\description{
Création de la matrice de vulnérabilité pour les sites d'importance fauniques et floristiques considérés pour l'évaluation des effets cumulatifs
}
\details{
Cette fonction génère la matrice de vulnérabilité pour les habitats
}
\keyword{composante}
\keyword{habitat}
\keyword{valorisée}
\keyword{vulnérabilité}
