% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuln_habitat.R
\name{vuln_habitat}
\alias{vuln_habitat}
\title{Vulnérabilité des habitats aux stresseurs environnementaux}
\usage{
vuln_habitat()
}
\description{
Création de la matrice de vulnérabilité pour les habtiats considérés pour l'évaluation des effets cumulatifs
}
\details{
Cette fonction génère la matrice de vulnérabilité pour les habitats
}
\keyword{composante}
\keyword{habitat}
\keyword{valorisée}
\keyword{vulnérabilité}
