% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuln_mammiferes_marins.R
\name{vuln_mammiferes_marins}
\alias{vuln_mammiferes_marins}
\title{Vulnérabilité des mammifères marins aux stresseurs environnementaux}
\usage{
vuln_mammiferes_marins()
}
\description{
Création de la matrice de vulnérabilité pour les mammifères marins considérés pour l'évaluation des effets cumulatifs
}
\details{
Cette fonction génère la matrice de vulnérabilité pour les habitats
}
\keyword{composante}
\keyword{mammifères}
\keyword{marins}
\keyword{valorisée}
\keyword{vulnérabilité}
