% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuln_site.R
\name{vuln_site}
\alias{vuln_site}
\title{Vulnérabilité des sites d'intérêt culturels, patrimoniaux et archéologiques aux stresseurs environnementaux}
\usage{
vuln_site()
}
\description{
Création de la matrice de vulnérabilité pour les sites d'intérêt culturels, patrimoniaux et archéologiques considérés pour l'évaluation des effets cumulatifs
}
\details{
Cette fonction génère la matrice de vulnérabilité pour les habitats
}
\keyword{archéologiques}
\keyword{composante}
\keyword{culturels,}
\keyword{d'intérêt}
\keyword{et}
\keyword{patrimoniaux}
\keyword{sites}
\keyword{valorisée}
\keyword{vulnérabilité}
