% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_vulnerability_matrix.R
\name{vulnerability_matrix}
\alias{vulnerability_matrix}
\title{Vulnérabilité des composantes valorisées aux stresseurs environnementaux}
\usage{
vulnerability_matrix()
}
\description{
Création de la matrice de vulnérabilité pour l'ensemble des composantes valorisées et stresseurs considérés pour l'évaluation des effets cumulatifs
}
\details{
Cette fonction génère la matrice de vulnérabilité pour les habitats
}
\keyword{composante}
\keyword{environnementaux}
\keyword{stresseurs}
\keyword{valorisée}
\keyword{vulnérabilité}
