% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_cumulative_footprint.R
\name{cumulativeFootprint}
\alias{cumulativeFootprint}
\title{Cumulative footprint}
\usage{
cumulativeFootprint(dat, normaliser = FALSE)
}
\arguments{
\item{dat}{data.frame with stressors intensity}

\item{normaliser}{logical, whether to normalize footprint}
}
\description{
Function to evaluate cumulative footprint of stressors
}
\details{
This function uses fisheries data to evaluate the intensity of fishing over a regular grid. If \code{normaliser} is TRUE, then the cumulative footprint is divided by the number of cumulated stressors.
}
\keyword{cumulative}
\keyword{footprint}
