% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_cv_habitat.R
\name{cv_habitat}
\alias{cv_habitat}
\title{Habitats: Données sur les habitats}
\usage{
cv_habitat()
}
\description{
Intégration des données utilisées pour caractériser la composante valorisée d'intégrité des habitats
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{valorisée}
