% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data0051-ballast.R
\name{fig_data0051_ballast}
\alias{fig_data0051_ballast}
\title{Figure of ballast exchange waters - data0051}
\usage{
fig_data0051_ballast()
}
\description{
Figure of ballast exchange waters - data0051
}
