% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0005-milieu_cotier.R
\name{get_data0005}
\alias{get_data0005}
\title{Data 0005 : Milieux humides cartographie détaillée}
\source{
https://www.donneesquebec.ca/recherche/dataset/eafec419-d67d-449e-a157-d22230314d36
}
\usage{
get_data0005()
}
\description{
Cartographie détaillée des milieux humides pour les basses-terres du Saint-Laurent, la plaine du Lac Saint-Jean et d'autres secteurs au sud du Québec
}
\details{
Cette fonction télécharge et formatte les données
}
\keyword{composante}
\keyword{habitat}
\keyword{humide}
\keyword{milieu}
\keyword{valorisée}
