% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0019-dragage.R
\name{get_data0019}
\alias{get_data0019}
\title{Data 0019 : Sites de dragage et de dépôt}
\source{
ECCC 2020. Bilan des activités de dragage dans l’Estuaire du Saint-Laurent de 2009 à 2016. Document de travail.
}
\usage{
get_data0019()
}
\description{
Bilan des activités de dragage dans l'estuaire du Saint-Laurent
}
\details{
Cette fonction formatte les données
}
\keyword{dragage}
\keyword{stresseurs}
