% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0029-marais.R
\name{get_data0029}
\alias{get_data0029}
\title{Data 0029 :}
\source{

}
\usage{
get_data0029()
}
\description{
Data 0029 :
}
\details{
Cette fonction formatte les données
}
\keyword{composante}
\keyword{marais}
\keyword{valorisée}
