% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0049-port_contrecoeur.R
\name{get_data0049}
\alias{get_data0049}
\title{Data 0049 : Dragage prévu Port de Montréal à Contrecoeur}
\source{
Ministère de l'Environnement et Lutte Contre les Changement Climatiques (2019) Rapport d’analyse environnementale Programme décennal de dragage d’entretien des installations portuaires de Port-Alfred sur le territoire de la ville de Saguenay. Direction générale de l'évaluation environnementale et stratégique. Direction de l'évaluation environnementale des projets hydriques et industriels. Dossier 3211-02-299. 29p. + viii

Englobe (2016) Programme décennal de dragage d’entretien et réfection des quais – Installations portuaires Port-Alfred, La Baie, Québec. Étude d’impact sur l’environnement. Rapport final 045-P-0008779-0-01-291-EN-R-0200-00. 132p. + ix
}
\usage{
get_data0049()
}
\description{
Identification des activités de dragage prévus au Port de Montréal à Contrecoeur
}
\details{
Cette fonction formatte les données
}
\keyword{dragage}
\keyword{stresseurs}
