% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0069-dragage_essipit.R
\name{get_data0069}
\alias{get_data0069}
\title{Data 0069 : Dragage Première Nation des Innus d'Essipit}
\source{
Bourchard D. (communication personnelle 2022) Position et volume dragués en 2001, 2011 et 2016 pour la Première Nation des Innus d'Essipit.
}
\usage{
get_data0069()
}
\description{
Site de dragage et de dépôt pour la Première Nation des Innus d'Essipit
}
\details{
Cette fonction formatte les données
}
\keyword{dragage}
