% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataXXXX-Conservation_measures.R,
%   R/dataXXXX-MPA.R, R/dataXXXX-Qualite_eau.R, R/dataXXXX-aires_protogees.R,
%   R/dataXXXX-depot_surface.R, R/dataXXXX-habitat_floristique.R,
%   R/dataXXXX-port.R, R/dataXXXX-reseay_hydrographique.R,
%   R/dataXXXX-suivi_bacteriologique.R
\name{get_dataXXXX}
\alias{get_dataXXXX}
\title{Data XXXX : Other Effective Area-Based Conservation Measures}
\source{
https://open.canada.ca/data/en/dataset/44769543-7a23-4991-a53f-c2cf7c7a946f

https://open.canada.ca/data/en/dataset/a1e18963-25dd-4219-a33f-1a38c4971250

https://www.donneesquebec.ca/recherche/dataset/suivi-physicochimique-des-rivieres-et-du-fleuve#

https://www.donneesquebec.ca/recherche/dataset/aires-protegees-au-quebec

https://www.donneesquebec.ca/recherche/dataset/depots-de-surface

https://www.donneesquebec.ca/recherche/dataset/habitats-floristiques

https://ouvert.canada.ca/data/fr/dataset/dcf460ae-c27f-4721-bec4-6e4467a5a3b6

https://www.donneesquebec.ca/recherche/dataset/grhq

https://www.donneesquebec.ca/recherche/dataset/suivi-bacteriologique-fleuve-sites-potentiels-baignade
}
\usage{
get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()
}
\description{
Oceans Act Marine Protected Areas

Oceans Act Marine Protected Areas
}
\details{
This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data
}
\keyword{}
