% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0040-MPA.R, R/data0041-port.R,
%   R/dataXXXX-Qualite_eau.R, R/dataXXXX-depot_surface.R,
%   R/dataXXXX-reseay_hydrographique.R, R/dataXXXX-suivi_bacteriologique.R
\name{get_dataXXXX}
\alias{get_dataXXXX}
\title{Data XXXX : Oceans Act Marine Protected Areas}
\source{
https://open.canada.ca/data/en/dataset/a1e18963-25dd-4219-a33f-1a38c4971250

https://ouvert.canada.ca/data/fr/dataset/dcf460ae-c27f-4721-bec4-6e4467a5a3b6

https://www.donneesquebec.ca/recherche/dataset/suivi-physicochimique-des-rivieres-et-du-fleuve#

https://www.donneesquebec.ca/recherche/dataset/depots-de-surface

https://www.donneesquebec.ca/recherche/dataset/grhq

https://www.donneesquebec.ca/recherche/dataset/suivi-bacteriologique-fleuve-sites-potentiels-baignade
}
\usage{
get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()
}
\description{
Oceans Act Marine Protected Areas
}
\details{
This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data
}
\keyword{}
