% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_plot_ceanav.R
\name{plot_ceanav}
\alias{plot_ceanav}
\alias{plot_ceanav.sf}
\title{plot ceanav data maps}
\usage{
plot_ceanav(dat, ...)

\method{plot_ceanav}{sf}(
  dat,
  main = NULL,
  type = NULL,
  subtitle = NULL,
  unit_data = NULL,
  references = NULL,
  city = TRUE,
  zones_NA = NULL,
  ...
)
}
\arguments{
\item{dat}{object of class sf}

\item{...}{further specifications, see \link{plot} and details.}

\item{main}{main title}

\item{type}{secondary title / data type (see metadata files)}

\item{subtitle}{subtitle}

\item{unit_data}{units of data}

\item{references}{data citation used for integrated data}

\item{city}{logical, if TRUE name of cities are added to graph}

\item{zones_NA}{character, name of zones for which data is not available. Data is available as a basemap in \code{./data/data-basemap/zonesNA.geojson}.}

\item{minUp}{numeric, minimum upper side to write as a function of bbox extent (legend)}
}
\description{
base plot functions for ceanav project
}
\examples{
 load_integrated("navigation")
 plot_ceanav(navigation[,10], "Navigation", "Recherche gouvernementale")

}
