% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_data_description.R, R/rep_data_summary.R
\name{rep_data_summary}
\alias{rep_data_summary}
\title{Fiches descriptives des données}
\usage{
rep_data_summary()

rep_data_summary()
}
\arguments{
\item{data_id}{\code{character} id of data to import in R session with format \code{dataXXXX}}
}
\description{
Fonction utilisée pour générer une fiches descriptive pour chaque données incluse au tableau récapitulatif des données à l'Annexe 1

Fonction utilisée pour générer un tableau récapitulatif des données présentées dans le rapport.
}
\details{
Cette fonction permet de générer une fiche descriptive pour la base de données sélectionnée

Cette fonction permet de générer le tableau récapitulatif des données utilisées pour le rapport
}
\examples{
rep_data_description("data0001")
rep_data_summary()
}
\keyword{descriptive}
\keyword{fiche}
\keyword{récapitulatif}
\keyword{tableau}
